requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
bcrypt
python-multipart
python-jose[cryptography]
passlib[bcrypt]
